src/test/java/org/actech/smart/trader/sync/stock/resolver/StockDetailHtmlResolverTest.java
package org.actech.smart.trader.sync.stock.resolver;

import org.actech.smart.trader.Application;
import org.actech.smart.trader.UnitTestConfiguration;
import org.actech.smart.trader.sync.stock.entity.StockFundTrack;
import org.jsoup.Jsoup;
import org.jsoup.nodes.Document;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.core.io.DefaultResourceLoader;
import org.springframework.core.io.Resource;
import org.springframework.core.io.ResourceLoader;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.junit4.SpringRunner;

import java.io.IOException;
import java.util.Collection;

import static org.hamcrest.core.Is.*;
import static org.hamcrest.core.IsNull.*;
import static org.junit.Assert.*;

/**
 * Created by paul on 2018/3/15.
 */
@RunWith(SpringRunner.class)
@SpringBootTest(classes = {Application.class, UnitTestConfiguration.class})
@ActiveProfiles("dev")
public class StockDetailHtmlResolverTest {
    @Autowired
    private StockDetailHtmlResolver parser;

    @Test
    public void shouldParseStockDetail() throws IOException {
        ResourceLoader resourceLoader = new DefaultResourceLoader();
        Resource resource = resourceLoader.getResource("classpath:海航基础(停牌).htm");
        Document document = Jsoup.parse(resource.getFile(), "GB2312");
        Collection<StockFundTrack> result = parser.resolve(document);
        assertThat(result, is(notNullValue()));
    }
}